Adjust Pokemon's height and weight properties type

+++ Pokemon/API/Pokemon.swift
@@ -9,13 +9,13 @@ import Foundation
 
 public struct Pokemon: Hashable {
     public let name: String
-    public let height: Int
-    public let weight: Int
+    public let height: Double
+    public let weight: Double
     public let id: Int
     public let type: String
     public let spritesImage: String
     
-    public init(name: String, height: Int, weight: Int, id: Int, type: String, spritesImage: String) {
+    public init(name: String, height: Double, weight: Double, id: Int, type: String, spritesImage: String) {
         self.name = name
         self.height = height
         self.weight = weight
